sheepdog: don't lock VDI when the driver is used via qemu-img or qemu-io
Recently, sheepdog updated its protocol and locking of VDI is
revived. This patch updates the locking request method of qemu-img and
qemu-io. One of their main usage is for taking snapshot of running
VDI. If they try to lock the VDI, it must conflict with running
QEMU. So this patch lets the commands not issue locking request during
opening VDI.

Cc: Jinzhi Chen <nxtjinzhi@gmail.com>
Cc: Valerio Pachera <sirio81@gmail.com>
Cc: Kevin Wolf <kwolf@redhat.com>
Cc: Stefan Hajnoczi <stefanha@redhat.com>
Cc: Liu Yuan <namei.unix@gmail.com>
Cc: Markus Armbruster <armbru@redhat.com>
Tested-by: Xiaolong Xu <nxtxiaolong@gmail.com>
Signed-off-by: Hitoshi Mitake <mitake.hitoshi@lab.ntt.co.jp>