sheepdog: add a function for removing objects before deleting VDIs
Though sheepdog is a distributed file system with minimal metadata
transaction, inode update must be serialized in the client
side, not in server side (sheep daemon). Currently, the sheepdog
driver serializes inode update in a correct manner for COW
requests. However, it should also be done in deletion process
(e.g. qemu-img snapshot -d).

Therefore this patch adds a new function remove_objects() to
the driver. The function removes objects with with filling zero in
inode->data_vdi_id[]. It

Signed-off-by: Hitoshi Mitake <mitake.hitoshi@lab.ntt.co.jp>